fix comment

+++ pkg/domain/plan_replayer.go
@@ -31,6 +31,7 @@ import (
 	domain_metrics "github.com/pingcap/tidb/pkg/domain/metrics"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/metrics"
+	"github.com/pingcap/tidb/pkg/objstore"
 	"github.com/pingcap/tidb/pkg/objstore/storeapi"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/terror"
@@ -94,6 +95,7 @@ func (p *dumpFileGcChecker) gcDumpFilesByPath(ctx context.Context, path string, 
 		logutil.BgLogger().Warn("get global ext storage failed", zap.String("category", "dumpFileGcChecker"), zap.Error(err))
 		return
 	}
+	deleteFile := strings.HasPrefix(storage.URI(), objstore.LocalURIPrefix)
 	opt := &storeapi.WalkOption{
 		SubDir: path,
 	}
@@ -113,12 +115,14 @@ func (p *dumpFileGcChecker) gcDumpFilesByPath(ctx context.Context, path string, 
 			canGC = !createTime.After(gcTargetTimeDefault)
 		}
 		if canGC {
-			err := storage.DeleteFile(ctx, fileName)
-			if err != nil {
-				logutil.BgLogger().Warn("remove file failed", zap.String("category", "dumpFileGcChecker"), zap.Error(err), zap.String("filename", fileName))
-				return nil
+			if deleteFile {
+				err := storage.DeleteFile(ctx, fileName)
+				if err != nil {
+					logutil.BgLogger().Warn("remove file failed", zap.String("category", "dumpFileGcChecker"), zap.Error(err), zap.String("filename", fileName))
+					return nil
+				}
+				logutil.BgLogger().Info("dumpFileGcChecker successful", zap.String("filename", fileName))
 			}
-			logutil.BgLogger().Info("dumpFileGcChecker successful", zap.String("filename", fileName))
 			if isPlanReplayer && p.sctx != nil {
 				deletePlanReplayerStatus(ctx, p.sctx, baseName)
 				p.planReplayerTaskStatus.clearFinishedTask()
